simple fixes
git-svn-id: svn://svn.macaulay2.com/Macaulay2/trunk@5032 a8012f80-5f13-0410-8dc0-eb77bd940c0d

+++ M2/Macaulay2/e/f4/F4.cpp
@@ -349,7 +349,7 @@ void F4GB::make_matrix()
 
   // DEBUGGING:
   fprintf(stderr, "--matrix--%ld by %ld\n", 
-	  mat->rows.size(), mat->columns.size());
+	  (long)mat->rows.size(), (long)mat->columns.size());
 
   //  show_row_info();
   //  show_column_info();
